Replace pew workon with pew in
Instead of requiring the presence of an 'inve' script in the virtual
environment's bin directory, just embed the code that would be in that
file, and run it in our own process.

Additionally, since the 'in' command would have so much in common with
the 'workon' command, just remove the 'workon' command entirely in favor
of 'in' with no arguments.

Use a temp_environ context manager to make working with inve something
that could potentially be done from an external script without having to
nuke the process. I don't expect it to be a typical use-case, but it's
easy enough to keep support for.

As is the case for much of this work, large portions of this code are
directly ripped from https://github.com/berdario/pew. Additionally, some
of this is taken from his branch merging in inve to run directly, which
is the same approach that I'm now using here.