Add --relative option to `show`
Add the ability to resolve the virtual environment to a path that is
relative to the WORKON_HOME. This allows for making a useful prompt
that successfully deals with namespaced virtual environments.

For instance, I can create a virtual environment called "test/me".

When I do 'pew ls', it will show 'test/' in the list. I can then do
'pew ls test', and it will show 'me'. I can use that virtualenv
by doing 'pew in test/me'.

'pew show' would show the full path to that virtualenv by default.
Prompt scripts usually take the highest folder from the VIRTUAL_ENV
environment variable, and use that. However, that doesn't account
for the namespace (folder) that it is in (test).

Now 'pew show --relative' will attempt to resolve the path based on
the WORKON_HOME, and would show 'test/me'. I can use this in my
bash prompt like this:

    if [ -n "`pew show --relative`" ]; then
        export PS1="(`pew show --relative`)$PS1"
    fi