Add ability to associate a project directory
The project directory, which the path to will be stored in a file
called '.profile' at the root of the virtual envronment, is
automatically entered when using the 'in' command.

Additionally, 'new' gets the options '--project' and '-a' ('-p' is
already taken) to set the project directory on environment creation.